# openapi.yml
openapi: 3.0.0
info:
  title: DroneDepot Public API v1
  description: Public API for DroneDepot front-end
  version: 1.0.0
  contact:
    name: DroneDepot Support
    email: ops@drone-depot.ai
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.drone-depot.ai
    description: Production server
paths:
...
  /v1/public/job-request:
    post:
      summary: Submit a job request
      description: Submit a new job request for aerial photography services
      tags:
        - Job Requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '200':
          description: Job request submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  jobNumber:
                    type: string
                    example: "DD-250808-0123"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Validation failed"
                  details:
                    type: array
                    items:
                      type: object
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Too many job requests. Please try again later."
components:
  schemas:
...
    JobRequest:
      type: object
      required:
        - fullName
        - email
        - industry
        - location
        - agreeToTerms
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1-555-123-4567"
        company:
          type: string
          example: "Acme Real Estate"
        industry:
          type: string
          enum: [real_estate, construction, events, marketing, other]
          example: "real_estate"
        location:
          type: object
          required:
            - addressLine1
            - city
            - state
            - postalCode
          properties:
            addressLine1:
              type: string
              example: "123 Main St"
            city:
              type: string
              example: "Austin"
            state:
              type: string
              example: "TX"
            postalCode:
              type: string
              example: "78701"
            lat:
              type: number
              example: 30.2672
            lng:
              type: number
              example: -97.7431
        dateWindow:
          type: object
          properties:
            start:
              type: string
              format: date
              example: "2024-02-01"
            end:
              type: string
              format: date
              example: "2024-02-15"
        budgetBand:
          type: string
          enum: [under_500, 500_1000, 1000_2500, 2500_plus]
          example: "1000_2500"
        mustHaveShots:
          type: string
          example: "Aerial view of the entire property"
        referral:
          type: string
          example: "Google Search"
        agreeToTerms:
          type: boolean
          example: true
        captchaToken:
          type: string
          example: "hcaptcha_token_here"
        _website:
          type: string
          description: Honeypot field - should be empty